Fix subtree movement on MySQL.
The problem is that if a SET has multiple values in MySQL, column changes in the first clause affect the subsequent CASE statements.

In other words, if we first change left_field and right_field, and then subsequently want to set tree_depth via a CASE(), the evaluation of that CASE will be affected by the left/right changes.

As part of this change, I also fixed the ORM change detection for moves, which can now understand both the relationship and the id field changing.